修复一些奇怪的命名空间引用 移除冗余的package

+++ AxibugEmuOnline.Client/Assets/VirtualNes.Core/NES.cs
@@ -1,12 +1,7 @@
-using Codice.CM.Client.Differences;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Runtime.ConstrainedExecution;
-using System.Runtime.InteropServices.ComTypes;
-using System.Security.Cryptography;
 using System.Text;
-using UnityEngine;
 using VirtualNes.Core.Debug;
 
 namespace VirtualNes.Core
@@ -1228,7 +1223,7 @@ namespace VirtualNes.Core
                 return;
             }
 
-            if ((m_TapeCycles -= (double)cycles) > 0)
+            if ((m_TapeCycles -= cycles) > 0)
                 return;
 
             m_TapeCycles += (nescfg.CpuClock / 32000.0);
